1.7e4 release cleanup
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@291 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/World.cxx
@@ -157,8 +157,8 @@ TeamColor		World::whoseBase(const float* pos) const
   for (int i = 1; i < NumTeams; i++) {
     float nx = pos[0] - bases[i][0];
     float ny = pos[1] - bases[i][1];
-    float rx = cosf(atanf(ny/nx)-bases[i][3]) * sqrt((ny * ny) + (nx * nx));
-    float ry = sinf(atanf(ny/nx)-bases[i][3]) * sqrt((ny * ny) + (nx * nx));
+    float rx = (float) (cosf(atanf(ny/nx)-bases[i][3]) * sqrt((ny * ny) + (nx * nx)));
+    float ry = (float) (sinf(atanf(ny/nx)-bases[i][3]) * sqrt((ny * ny) + (nx * nx)));
     if(fabsf(rx) < bases[i][4] &&
        fabsf(ry) < bases[i][5]) {
       float nz = (bases[i][2] > 0) ? (bases[i][2] + 1) : 0;
